+++ src/components/Home/Body/Body.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import './Body.css';
-import prop1 from '../images/property1.jpg';
-import prop2 from '../images/property2.jpg';
-import prop3 from '../images/property3.jpg';
+import '../Body/Body.css';
+import prop1 from '../../../images/property1.jpg';
+import prop2 from '../../../images/property2.jpg';
+import prop3 from '../../../images/property3.jpg';
 
 function Body() {
   return (
