+++ src/components/Home/Home.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import Header from './Header';
-import Body from './Body';
-import Footer from './Footer';
+import Header from './Header/Header';
+import Body from './Body/Body';
+import Footer from './Footer/Footer.js';
 
 export default class Home extends React.Component {
   render() {
